refactor: keep EP runtime config backend-specific
