+++ torchtitan/distributed/minimal_async_ep/__init__.py
@@ -8,6 +8,7 @@ from torchtitan.distributed.minimal_async_ep.api import (
     combine_op,
     dispatch_op,
     init_buffer,
+    maybe_update_minimal_async_ep_config,
     MinimalAsyncEPDispatchMetadata,
 )
 
@@ -16,4 +17,5 @@ __all__ = [
     "combine_op",
     "dispatch_op",
     "init_buffer",
+    "maybe_update_minimal_async_ep_config",
 ]
