+++ torchtitan/distributed/minimal_async_ep/api.py
@@ -79,6 +79,84 @@ _buffer_key: (
 ) = None
 
 
+def maybe_update_minimal_async_ep_config(model_config: Any, config: Any) -> None:
+    """Validate and fill MinimalAsyncEP dispatcher configs from runtime config."""
+    from torchtitan.config import ParallelismConfig, TORCH_DTYPE_MAP
+    from torchtitan.distributed.activation_checkpoint import FullAC
+    from torchtitan.models.common.token_dispatcher import MinimalAsyncEPTokenDispatcher
+    from torchtitan.trainer import Trainer
+
+    assert hasattr(
+        config, "parallelism"
+    ), "config passed to update_from_config must provide a parallelism field."
+    parallelism = config.parallelism
+    assert isinstance(parallelism, ParallelismConfig), (
+        "config.parallelism must be a ParallelismConfig, got "
+        f"{type(parallelism).__name__}."
+    )
+
+    dispatcher_cfgs = []
+    for layer_cfg in model_config.layers:
+        moe_cfg = getattr(layer_cfg, "moe", None)
+        if moe_cfg is None:
+            continue
+        token_dispatcher_cfg = moe_cfg.routed_experts.token_dispatcher
+        if isinstance(token_dispatcher_cfg, MinimalAsyncEPTokenDispatcher.Config):
+            dispatcher_cfgs.append(token_dispatcher_cfg)
+
+    if not dispatcher_cfgs:
+        return
+
+    if parallelism.spmd_backend == "full_dtensor":
+        raise ValueError("MinimalAsyncEP does not support full_dtensor SPMD.")
+    if parallelism.expert_parallel_degree == 1:
+        raise ValueError(
+            "MinimalAsyncEPTokenDispatcher.Config requires expert parallelism "
+            "(expert_parallel_degree > 1)."
+        )
+    if parallelism.tensor_parallel_degree != 1:
+        raise ValueError(
+            "MinimalAsyncEP does not support tensor or sequence parallelism."
+        )
+    if parallelism.context_parallel_degree != 1:
+        raise ValueError("MinimalAsyncEP does not support context parallelism.")
+    if parallelism.pipeline_parallel_degree != 1:
+        raise ValueError("MinimalAsyncEP does not support pipeline parallelism.")
+    for num_experts in {cfg.num_experts for cfg in dispatcher_cfgs}:
+        if num_experts % parallelism.expert_parallel_degree != 0:
+            raise ValueError(
+                f"MinimalAsyncEP num_experts ({num_experts}) must be "
+                "divisible by expert_parallel_degree "
+                f"({parallelism.expert_parallel_degree})."
+            )
+
+    if not isinstance(config, Trainer.Config):
+        raise ValueError(
+            "MinimalAsyncEP requires a Trainer.Config-compatible runtime config "
+            "to set hidden_dim, tokens_per_rank, and dtype."
+        )
+
+    memory_policy = getattr(config.compile, "memory_policy", None)
+    if (
+        not isinstance(config.activation_checkpoint, FullAC.Config)
+        and memory_policy != "full"
+    ):
+        raise ValueError(
+            "MinimalAsyncEP requires full recompute: set "
+            "activation-checkpoint:full for eager training or "
+            "--compile.memory_policy full for graph_trainer."
+        )
+
+    for token_dispatcher_cfg in dispatcher_cfgs:
+        token_dispatcher_cfg.hidden_dim = model_config.dim
+        token_dispatcher_cfg.tokens_per_rank = (
+            config.training.local_batch_size * config.training.seq_len
+        )
+        token_dispatcher_cfg.dtype = TORCH_DTYPE_MAP[
+            config.training.mixed_precision_param
+        ]
+
+
 @dataclass
 class MinimalAsyncEPDispatchMetadata:
     """MinimalAsyncEP metadata from dispatch needed for combine.
@@ -952,4 +1030,5 @@ __all__ = [
     "combine_op",
     "dispatch_op",
     "init_buffer",
+    "maybe_update_minimal_async_ep_config",
 ]
