+++ torchtitan/models/common/decoder.py
@@ -11,6 +11,9 @@ from dataclasses import dataclass
 import torch
 from torch.nn.attention.flex_attention import _mask_mod_signature, and_masks, BlockMask
 
+from torchtitan.distributed.minimal_async_ep.api import (
+    maybe_update_minimal_async_ep_config,
+)
 from torchtitan.distributed.utils import is_in_batch_invariant_mode
 from torchtitan.models.common.attention import (
     AttentionMasksType,
@@ -27,9 +30,7 @@ from torchtitan.models.common.feed_forward import FeedForward
 from torchtitan.models.common.linear import Linear
 from torchtitan.models.common.moe import MoE
 from torchtitan.models.common.nn_modules import RMSNorm
-from torchtitan.models.common.token_dispatcher_config import (
-    update_ep_token_dispatcher_config,
-)
+from torchtitan.models.common.token_dispatcher import maybe_update_hybrid_ep_config
 from torchtitan.protocols.model import BaseModel
 from torchtitan.protocols.module import Module, ModuleDict
 
@@ -159,7 +160,32 @@ class Decoder(BaseModel):
                         f"n_kv_heads ({n_kv_heads})."
                     )
 
-            update_ep_token_dispatcher_config(self, config)
+            for layer_cfg in self.layers:
+                if layer_cfg.moe is not None:
+                    from torchtitan.models.common.token_dispatcher import (
+                        DeepEPTokenDispatcher,
+                        HybridEPTokenDispatcher,
+                    )
+
+                    token_dispatcher_cfg = layer_cfg.moe.routed_experts.token_dispatcher
+                    if (
+                        isinstance(
+                            token_dispatcher_cfg,
+                            (
+                                DeepEPTokenDispatcher.Config,
+                                HybridEPTokenDispatcher.Config,
+                            ),
+                        )
+                        and parallelism.expert_parallel_degree == 1
+                    ):
+                        raise ValueError(
+                            f"{type(token_dispatcher_cfg).__qualname__} "
+                            "requires expert parallelism "
+                            "(expert_parallel_degree > 1)."
+                        )
+
+            maybe_update_hybrid_ep_config(self, config)
+            maybe_update_minimal_async_ep_config(self, config)
 
             # NOTE: Inference-only callers such as the RL generator skip
             # training.seq_len sync. Generated sequence length is not known
